requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-polymorphic==3.0.0
idna==2.10
pycryptodome==3.9.9
pytz==2020.5
rave-python==1.2.9
requests==2.25.1
rope==0.18.0
sqlparse==0.4.1
urllib3==1.26.3
kafka-python==2.0.2
redis==3.5.3
djangorestframework==3.12.2 